+++ src/Collection/ObjectCollection.php
@@ -45,23 +45,6 @@
 
 
     /**
-     * @deprecated
-     * @see add
-     *
-     * Add one item to the end of collection
-     * This item is accessible via `$collection->getLast();`
-     *
-     * @param ElementFinder $item
-     * @return self
-     */
-    public function append(ElementFinder $item): self {
-      trigger_error('Deprecated. See add method', E_USER_DEPRECATED);
-      $this->items[] = $item;
-      return $this;
-    }
-
-
-    /**
      * Return last item from collection
      *
      * @return null|ElementFinder